+++ nerdlets/home/index.js
@@ -26,36 +26,48 @@ import {
 } from '../../src/components';
 import {
   useFlowLoader,
-  useFlowWriter,
   useFetchUser,
   useReadUserPreferences,
 } from '../../src/hooks';
+import { exportFlowDoc, flowDocument } from '../../src/utils';
 import { MODES, UI_CONTENT } from '../../src/constants';
 import { AppContext, SidebarProvider } from '../../src/contexts';
 
-const createFlowButtonAttributes = {
-  label: UI_CONTENT.GLOBAL.BUTTON_LABEL_CREATE_FLOW,
-  type: Button.TYPE.PRIMARY,
-  iconType: Icon.TYPE.DATAVIZ__DATAVIZ__SERVICE_MAP_CHART,
-};
-
-const editButtonAttributes = {
-  label: UI_CONTENT.GLOBAL.BUTTON_LABEL_EDIT_MODE,
-  type: Button.TYPE.PRIMARY,
-  iconType: Icon.TYPE.INTERFACE__OPERATIONS__EDIT,
-};
-
-const editButtonFlowSettingsAttributes = {
-  type: Button.TYPE.PRIMARY,
-  iconType: Icon.TYPE.INTERFACE__OPERATIONS__CONFIGURE,
+const ACTION_BTN_ATTRIBS = {
+  CREATE_FLOW: {
+    type: Button.TYPE.PRIMARY,
+    label: UI_CONTENT.GLOBAL.BUTTON_LABEL_CREATE_FLOW,
+    iconType: Icon.TYPE.DATAVIZ__DATAVIZ__SERVICE_MAP_CHART,
+  },
+  EDIT_FLOW: {
+    type: Button.TYPE.SECONDARY,
+    iconType: Icon.TYPE.INTERFACE__OPERATIONS__EDIT,
+    hint: UI_CONTENT.GLOBAL.BUTTON_LABEL_EDIT_MODE,
+  },
+  EDIT_FLOW_SETTINGS: {
+    type: Button.TYPE.TERTIARY,
+    label: UI_CONTENT.GLOBAL.BUTTON_LABEL_EDIT_SETTINGS,
+  },
+  EXPORT_FLOW: {
+    type: Button.TYPE.TERTIARY,
+    label: UI_CONTENT.GLOBAL.BUTTON_LABEL_EXPORT_FLOW,
+  },
+  AUDIT_LOG: {
+    type: Button.TYPE.TERTIARY,
+    label: UI_CONTENT.GLOBAL.BUTTON_LABEL_AUDIT_LOG,
+  },
 };
 
 const HomeNerdlet = () => {
   const [app, setApp] = useState({});
   const [mode, setMode] = useState(MODES.INLINE);
+  const [prevNonEditMode, setPrevNonEditMode] = useState(MODES.INLINE);
   const [flows, setFlows] = useState([]);
   const [currentFlowId, setCurrentFlowId] = useState();
+  const [isAuditLogShown, setisAuditLogShown] = useState(false);
   const [editFlowSettings, setEditFlowSettings] = useState(false);
+  const [transitionToFlow, setTransitionToFlow] = useState(false);
+  const [transitionIntervalId, setTransitionIntervalId] = useState();
   const { accountId } = useContext(PlatformStateContext);
   const [nerdletState, setNerdletState] = useNerdletState();
   const { user } = useFetchUser();
@@ -67,8 +79,13 @@ const HomeNerdlet = () => {
     loading: flowsLoading,
     refetch: flowsRefetch,
   } = useFlowLoader({ accountId });
-  const flowWriter = useFlowWriter({ accountId, user });
   const { data: accounts = [] } = useAccountsQuery();
+
+  useEffect(() => {
+    return () => {
+      if (transitionIntervalId) clearInterval(transitionIntervalId);
+    };
+  }, [transitionIntervalId]);
 
   useEffect(
     () =>
@@ -84,41 +101,59 @@ const HomeNerdlet = () => {
   );
 
   useEffect(() => {
+    const defaultButtons = currentFlowId
+      ? [
+          {
+            ...ACTION_BTN_ATTRIBS.EDIT_FLOW_SETTINGS,
+            onClick: () => setEditFlowSettings(true),
+          },
+        ]
+      : [
+          {
+            ...ACTION_BTN_ATTRIBS.CREATE_FLOW,
+            onClick: newFlowHandler,
+          },
+        ];
+
+    const actionControlButtons =
+      currentFlowId && mode !== MODES.EDIT
+        ? [
+            {
+              ...ACTION_BTN_ATTRIBS.EDIT_FLOW,
+              onClick: () => changeMode(MODES.EDIT),
+            },
+            {
+              ...ACTION_BTN_ATTRIBS.EXPORT_FLOW,
+              onClick: () => exportFlowDoc(flows, currentFlowId),
+            },
+            {
+              ...ACTION_BTN_ATTRIBS.AUDIT_LOG,
+              onClick: () => setisAuditLogShown(true),
+            },
+            ...defaultButtons,
+            {
+              ...ACTION_BTN_ATTRIBS.CREATE_FLOW,
+              onClick: newFlowHandler,
+            },
+          ]
+        : [...defaultButtons];
+
     nerdlet.setConfig({
       accountPicker: true,
       actionControls: true,
-      actionControlButtons: currentFlowId
-        ? [
-            {
-              ...createFlowButtonAttributes,
-              onClick: newFlowHandler,
-            },
-            {
-              ...editButtonFlowSettingsAttributes,
-              onClick: () => setEditFlowSettings(true),
-            },
-            {
-              ...editButtonAttributes,
-              onClick: () => setMode(MODES.EDIT),
-            },
-          ]
-        : [
-            {
-              ...createFlowButtonAttributes,
-              onClick: newFlowHandler,
-            },
-          ],
+      actionControlButtons,
       headerType: nerdlet.HEADER_TYPE.CUSTOM,
       headerTitle: 'Project Hedgehog 🦔',
     });
-  }, [user, newFlowHandler, currentFlowId, editFlowSettings]);
+  }, [currentFlowId, flows, mode]);
 
   useEffect(() => setCurrentFlowId(nerdletState.flow?.id), [nerdletState.flow]);
 
-  useEffect(
-    () => setMode(nerdletState.mode || MODES.INLINE),
-    [nerdletState.mode]
-  );
+  useEffect(() => {
+    const newMode = nerdletState.mode || MODES.INLINE;
+    if (newMode !== MODES.EDIT) setPrevNonEditMode(newMode);
+    setMode(newMode);
+  }, [nerdletState.mode]);
 
   useEffect(() => {
     if (nerdletState.refreshFlows) {
@@ -139,9 +174,9 @@ const HomeNerdlet = () => {
     });
 
   const changeMode = useCallback(
-    (mode = MODES.INLINE) =>
+    (newMode = MODES.INLINE) =>
       setNerdletState({
-        mode,
+        mode: newMode,
       }),
     []
   );
@@ -160,21 +195,22 @@ const HomeNerdlet = () => {
     []
   );
 
+  const transitionToMode = useCallback((newMode) => {
+    setTransitionToFlow(true);
+    setTransitionIntervalId(
+      setInterval(() => {
+        setTransitionToFlow(false);
+      }, 3000)
+    );
+    if (newMode) changeMode(newMode);
+  }, []);
+
+  const auditLogCloseHandler = () => setisAuditLogShown(false);
+
   const currentFlowDoc = useMemo(
-    () =>
-      currentFlowId
-        ? (flows || []).find(({ id }) => id === currentFlowId)?.document
-        : null,
+    () => (currentFlowId ? flowDocument(flows, currentFlowId) : null),
     [currentFlowId, flows]
   );
-
-  useEffect(() => {
-    const { nerdStorageWriteDocument: { id } = {} } = flowWriter?.data || {};
-    if (id) {
-      setMode(MODES.EDIT);
-      flowClickHandler(id);
-    }
-  }, [flowWriter.data]);
 
   const currentView = useMemo(() => {
     if (
@@ -193,8 +229,12 @@ const HomeNerdlet = () => {
               onClose={backToFlowsHandler}
               mode={mode}
               setMode={changeMode}
+              prevNonEditMode={prevNonEditMode}
               flows={flows}
               onSelectFlow={flowClickHandler}
+              onTransition={transitionToMode}
+              isAuditLogShown={isAuditLogShown}
+              onAuditLogClose={auditLogCloseHandler}
               editFlowSettings={editFlowSettings}
               setEditFlowSettings={setEditFlowSettings}
             />
@@ -218,10 +258,15 @@ const HomeNerdlet = () => {
     accountId,
     mode,
     flowClickHandler,
+    isAuditLogShown,
     editFlowSettings,
   ]);
 
-  return <div className="container">{currentView}</div>;
+  return transitionToFlow ? (
+    <Spinner />
+  ) : (
+    <div className="container">{currentView}</div>
+  );
 };
 
 export default HomeNerdlet;