+++ src/components/flow/header.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { memo, useCallback, useState } from 'react';
 import PropTypes from 'prop-types';
 
 import {
@@ -20,27 +20,98 @@ import { MODES, UI_CONTENT } from '../../constants';
 const FlowHeader = ({
   name = 'Flow',
   imageUrl,
+  isPreview,
+  onPreview,
   onUpdate,
+  onDiscard,
+  onPersist,
   onClose,
   mode = MODES.INLINE,
   setMode = () => null,
   flows = [],
   onSelectFlow = () => null,
-  onExportFlow = () => null,
   onDeleteFlow = () => null,
   lastSavedTimestamp,
-  resetLastSavedTimestamp = () => null,
 }) => {
   const [imageModalHidden, setImageModalHidden] = useState(true);
-
-  const doneEditingHandler = () => {
-    resetLastSavedTimestamp();
-    setMode(MODES.INLINE);
-  };
 
   const capitalize = useCallback((word) =>
     word.replace(/\b\w/g, (l) => l.toUpperCase())
   );
+
+  const EditButtons = memo(
+    () => (
+      <>
+        <Button
+          type={Button.TYPE.TERTIARY}
+          iconType={Button.ICON_TYPE.INTERFACE__OPERATIONS__CLOSE}
+          sizeType={Button.SIZE_TYPE.SMALL}
+          onClick={onDiscard}
+        >
+          {UI_CONTENT.GLOBAL.BUTTON_LABEL_EDIT_DISCARD}
+        </Button>
+        {isPreview ? (
+          <Button
+            type={Button.TYPE.SECONDARY}
+            iconType={Button.ICON_TYPE.INTERFACE__OPERATIONS__EDIT}
+            sizeType={Button.SIZE_TYPE.SMALL}
+            onClick={onPreview}
+          >
+            {UI_CONTENT.GLOBAL.BUTTON_LABEL_PREVIEW_EXIT}
+          </Button>
+        ) : (
+          <Button
+            type={Button.TYPE.SECONDARY}
+            iconType={Button.ICON_TYPE.INTERFACE__OPERATIONS__SHOW}
+            sizeType={Button.SIZE_TYPE.SMALL}
+            onClick={onPreview}
+          >
+            {UI_CONTENT.GLOBAL.BUTTON_LABEL_PREVIEW}
+          </Button>
+        )}
+        <Button
+          type={Button.TYPE.PRIMARY}
+          iconType={Button.ICON_TYPE.INTERFACE__SIGN__CHECKMARK}
+          sizeType={Button.SIZE_TYPE.SMALL}
+          onClick={onPersist}
+        >
+          {UI_CONTENT.GLOBAL.BUTTON_LABEL_EDIT_PERSIST}
+        </Button>
+      </>
+    ),
+    [isPreview]
+  );
+  EditButtons.displayName = 'EditButtons';
+
+  const ViewModeSegmentedControl = memo(
+    () => (
+      <SegmentedControl
+        className="view-mode"
+        type={
+          isPreview
+            ? SegmentedControl.TYPE.ICONS_ONLY
+            : SegmentedControl.TYPE.NORMAL
+        }
+        value={mode}
+        onChange={(_, value) => setMode(value)}
+      >
+        <SegmentedControlItem
+          label={capitalize(MODES.STACKED)}
+          value={MODES.STACKED}
+          iconType={
+            SegmentedControlItem.ICON_TYPE.DATAVIZ__DATAVIZ__TABLE_CHART
+          }
+        />
+        <SegmentedControlItem
+          label={capitalize(MODES.INLINE)}
+          value={MODES.INLINE}
+          iconType={SegmentedControlItem.ICON_TYPE.INTERFACE__VIEW__LIST_VIEW}
+        />
+      </SegmentedControl>
+    ),
+    [isPreview]
+  );
+  ViewModeSegmentedControl.displayName = 'ViewModeSegmentedControl';
 
   return mode === MODES.EDIT ? (
     <div className="flow-header">
@@ -79,14 +150,7 @@ const FlowHeader = ({
         onClose={() => setImageModalHidden(true)}
       />
       <div className="flow-header-actions">
-        <Button
-          type={Button.TYPE.PRIMARY}
-          iconType={Icon.TYPE.INTERFACE__SIGN__CHECKMARK}
-          sizeType={Button.SIZE_TYPE.SMALL}
-          onClick={doneEditingHandler}
-        >
-          {UI_CONTENT.GLOBAL.BUTTON_LABEL_EXIT_EDIT_MODE}
-        </Button>
+        <EditButtons />
         <Button
           type={Button.TYPE.DESTRUCTIVE}
           iconType={Icon.TYPE.INTERFACE__OPERATIONS__TRASH}
@@ -122,32 +186,14 @@ const FlowHeader = ({
         </PopoverBody>
       </Popover>
       <div className="flow-header-actions">
-        <Button
-          type={Button.TYPE.PLAIN}
-          sizeType={Button.SIZE_TYPE.SMALL}
-          iconType={Button.ICON_TYPE.INTERFACE__OPERATIONS__DOWNLOAD}
-          onClick={onExportFlow}
-        >
-          {UI_CONTENT.FLOW.BUTTON_EXPORT}
-        </Button>
-        <SegmentedControl
-          className="view-mode"
-          value={mode}
-          onChange={(_, value) => setMode(value)}
-        >
-          <SegmentedControlItem
-            label={capitalize(MODES.STACKED)}
-            value={MODES.STACKED}
-            iconType={
-              SegmentedControlItem.ICON_TYPE.DATAVIZ__DATAVIZ__TABLE_CHART
-            }
-          />
-          <SegmentedControlItem
-            label={capitalize(MODES.INLINE)}
-            value={MODES.INLINE}
-            iconType={SegmentedControlItem.ICON_TYPE.INTERFACE__VIEW__LIST_VIEW}
-          />
-        </SegmentedControl>
+        {isPreview ? (
+          <>
+            <EditButtons />
+            <ViewModeSegmentedControl />
+          </>
+        ) : (
+          <ViewModeSegmentedControl />
+        )}
       </div>
     </div>
   );
@@ -156,16 +202,18 @@ const FlowHeader = ({
 FlowHeader.propTypes = {
   name: PropTypes.string,
   imageUrl: PropTypes.string,
+  isPreview: PropTypes.bool,
+  onPreview: PropTypes.func,
   onUpdate: PropTypes.func,
+  onDiscard: PropTypes.func,
+  onPersist: PropTypes.func,
   onClose: PropTypes.func,
   mode: PropTypes.oneOf(Object.values(MODES)),
   setMode: PropTypes.func,
   flows: PropTypes.array,
   onSelectFlow: PropTypes.func,
-  onExportFlow: PropTypes.func,
   onDeleteFlow: PropTypes.func,
   lastSavedTimestamp: PropTypes.number,
-  resetLastSavedTimestamp: PropTypes.func,
 };
 
 export default FlowHeader;