+++ src/components/index.js
@@ -20,3 +20,4 @@ export { default as Sidebar } from './sidebar';
 export { default as GetStarted } from './get-started';
 export { default as EditFlowSettingsModal } from './flow-settings-modal';
 export { default as Select } from './select';
+export { default as AuditLog } from './audit-log';